requirements.txt
appdirs==1.3.0
Jinja2==2.7.2
MarkupSafe==0.23
MySQL-python==1.2.5
PyYAML==3.11
Pygments==1.6
SQLAlchemy==0.9.4
Sphinx==1.2.2
docutils==0.11
py==1.4.20
tox==1.7.1
unittest-xml==0.2.2
unittest-xml-reporting==1.7.0
virtualenv==1.11.5
wsgiref==0.1.2
xmlunittest==0.2.0